refactor(exp): rename function_app.py to main.py
The exp service is a plain FastAPI app, not an Azure Functions app. Rename the entrypoint to main.py and update all references in Dockerfile, conftest.py, README.md, and the colocated test file.
